radio group working, distribute types
