+++ client/scripts/views/components/component_kit/cw_button.tsx
@@ -3,8 +3,18 @@
 import m from 'mithril';
 import 'components/component_kit/cw_button.scss';
 
-import { ComponentType, State, ButtonProps } from './types';
+import { ButtonType, ComponentType, State, StyleProps } from './types';
 import { getClasses } from './helpers';
+
+type ButtonStyleProps = {
+  buttonType: ButtonType;
+} & StyleProps;
+
+export type ButtonProps = {
+  styleProps: ButtonStyleProps;
+  label: string;
+  onclick: (e?: MouseEvent) => void;
+};
 
 export const CWButton: m.Component<ButtonProps, State> = {
   view: (vnode) => {
