+++ client/scripts/views/components/component_kit/cw_external_link.tsx
@@ -4,7 +4,18 @@ import m from 'mithril';
 import 'components/component_kit/cw_external_link.scss';
 
 import { ExternalLinkIcon } from './icons';
-import { ComponentType, ExternalLinkProps, State } from './types';
+import { ComponentType, State } from './types';
+
+export enum LinkType {
+  Button = 'button',
+  Inline = 'inline',
+}
+
+type ExternalLinkProps = {
+  label: string;
+  target: string;
+  linkType: LinkType;
+};
 
 // TODO: Graham 11/17/21 - Synchronize/reconcile against Mithril internal/external link helpers
 export const CWExternalLink: m.Component<ExternalLinkProps, State> = {
