+++ client/scripts/views/components/component_kit/cw_radio_button.tsx
@@ -5,21 +5,20 @@ import 'components/component_kit/cw_radio_button.scss';
 
 import { ComponentType, State } from './types';
 
-export const CWRadioButton: m.Component<
-  {
-    value: string;
-    label?: string;
-    checked: boolean;
-    groupName: string;
-    // TODO: Gabe 1/14/22 type onchange for real
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    onchange: (e?: any) => void;
-    className?: string;
-    // TODO: Gabe 1/14/22 disabled isn't used for anything
-    disabled?: boolean;
-  },
-  State
-> = {
+export type RadioButtonProps = {
+  value: string;
+  label?: string;
+  checked: boolean;
+  groupName: string;
+  // TODO: Gabe 1/14/22 type onchange for real
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  onchange: (e?: any) => void;
+  className?: string;
+  // TODO: Gabe 1/14/22 disabled isn't used for anything
+  disabled?: boolean;
+};
+
+export const CWRadioButton: m.Component<RadioButtonProps, State> = {
   view: (vnode) => {
     const { checked, value, label, groupName, onchange, className } =
       vnode.attrs;
