+++ client/styles/components/component_kit/buttons.scss
@@ -1,12 +1,5 @@
 @import 'client/styles/facelift';
 
-.RadioGroup {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-}
-
 .EngagementButton {
   display: flex;
   align-items: center;
