+++ client/styles/components/component_kit/cw_radio_group.scss
@@ -1,0 +1,6 @@
+.RadioGroup {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  width: 100%;
+}
